+++ internal/dinosql/query_test.go
@@ -4,8 +4,9 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/kyleconroy/dinosql/internal/postgres"
+
 	"github.com/google/go-cmp/cmp"
-	"github.com/kyleconroy/dinosql/postgres"
 )
 
 const testAliasSQL = `
