+++ internal/dinosql/testdata/funcs/math.sql
