+++ internal/dinosql/testdata/ondeck/schema/0001_city.sql
