+++ internal/dinosql/testdata/ondeck/schema/0002_venue.sql
