+++ internal/dinosql/testdata/ondeck/schema/0003_add_column.sql
